Create a minimal CWL, as a target for the conversion
