+++ demo/testdata/minimal-cwl.js
@@ -1,0 +1,23 @@
+export const STEPS = [
+    {
+        "name":"bedtobeddb",
+        "inputs":[
+            {
+                "meta":{ "global": true },
+                "name":"bedfile",
+                "source":[]
+            }
+        ],
+        "outputs":[
+            {
+                "meta":{ "global":true },
+                "name":"beddbfile",
+                "target":[
+                    {
+                        "name":"beddbfile"
+                    }
+                ]
+            }
+        ]
+    }
+];
